Reactor: один элемент за итерацию + корректные типы очередей
Реализован правильный reactor паттерн для dap_context_queue:
- Обработка только одного элемента за итерацию (атомарность)
- Re-signal для последующих элементов в очереди
- Возврат int (0/-1) вместо size_t для dap_context_queue_process

Обновлены все типы очередей:
- dap_stream_worker_t: queue_ch_io/send/input → dap_context_queue_t*
- Callbacks без dap_events_socket_t*, используют dap_worker_get_current()
- Все inter-worker функции (dap_stream_ch_pkt_write_inter, avrs_ch_pkt_send_*_inter)
  принимают dap_context_queue_t* вместо dap_events_socket_t*

Исправлены заголовочные файлы:
- Добавлено объявление dap_events_socket_event_signal() в dap_events_socket.h
- Добавлен include dap_context_queue.h в dap_context.c
- Добавлен include <inttypes.h> в dap_ring_buffer.c для PRIu64

Все модули DAP SDK компилируются без ошибок.

+++ io/include/dap_context_queue.h
@@ -71,11 +71,11 @@ void dap_context_queue_delete(dap_context_queue_t *a_queue);
 bool dap_context_queue_push(dap_context_queue_t *a_queue, void *a_item);
 
 /**
- * @brief Process all available items in queue (called by reactor)
+ * @brief Process one item from queue (called by reactor)
  * @param a_queue Queue
- * @return Number of items processed
+ * @return 0 on success, -1 on error
  */
-size_t dap_context_queue_process(dap_context_queue_t *a_queue);
+int dap_context_queue_process(dap_context_queue_t *a_queue);
 
 /**
  * @brief Get queue statistics
